# openapi.yaml
openapi: 3.0.3
info:
  title: TalentForge API Specifications
  description: Complete OpenAPI API specifications for the TalentForge performance-verified talent marketplace.
  version: 0.1.0-alpha
servers:
  - url: http://localhost:5000/api
    description: Local Dev Server
paths:
...
  /auth/register:
    post:
      summary: Register candidate account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
                - domain
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
                domain:
                  type: string
                  enum: [cse, ece]
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Payload validation failed
        '409':
          description: Email already in use
  /auth/login:
    post:
      summary: Login candidate session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Authenticated successfully. Returns JWT and refresh token.
        '401':
          description: Invalid email or password
  /auth/refresh:
    post:
      summary: Refresh JWT credentials token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token rotated successfully. Returns new JWT.
        '401':
          description: Invalid or expired refresh token
...
  /auth/me:
    get:
      summary: Retrieve candidate user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Candidate details retrieved
        '401':
          description: Unauthorized session
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT